rangecache: Add exact range match table test.
An exact range match is when we have incoming range requests that
correspond exactly with already stored key ranges in the range cache.

Note, in practice we intend to store the result of an HTTP range request
for video content. For the sake of visualizing and simpler logging, we
assume the associated data of any key range is the sequence of integers
from start to end, inclusive.