rangecache: Add eviction logic to Add.
On Add, if there is no cache hit, then before adding we loop until we have
enough available bytes to store the element's data.
A table test is added to test eviction, viz. "exact range match, evict
lru keyrange element".

Remark, we are counting the bytes based on our generated values ([]int).
We assume we are working on a 64-bit architecture where int is 64-bits.
For a more robust solution, we should use type switches to verify the
underlying type and count bytes.